+++ R/as-data-frame.R
@@ -1,0 +1,11 @@
+#' @export
+as.data.frame.ypr_population <- function(x, ...) {
+  x <- unclass(x)
+  as.data.frame(x)
+}
+
+#' @export
+as.data.frame.ypr_populations <- function(x, ...) {
+  x <- lapply(x, as.data.frame)
+  do.call("rbind", x)
+}